Module-B ML/requirements.txt
# Core ML and NLP
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0

# Vector Database and Search
faiss-cpu>=1.7.0

# Web Framework
flask>=2.2.0
flask-cors>=4.0.0

# Data Processing and Utils
pyyaml>=6.0
pickle5>=0.0.11; python_version < "3.8"

# Development and Testing
jupyter>=1.0.0
ipykernel>=6.15.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and Monitoring
python-dotenv>=0.19.0

# Optional: GPU support (uncomment if using GPU)
# faiss-gpu>=1.7.0
